+++ lib/codecept.js
@@ -46,7 +46,7 @@ class Codecept {
       global.locate = locator => require('./locator').build(locator);
 
       // BDD
-      const stepDefinitions = require('./interfaces/context');
+      const stepDefinitions = require('./interfaces/bdd');
       global.Given = stepDefinitions.Given;
       global.When = stepDefinitions.When;
       global.Then = stepDefinitions.Then;
