+++ lib/container.js
@@ -3,7 +3,6 @@ const fileExists = require('./utils').fileExists;
 const Translation = require('./translation');
 const MochaFactory = require('./mochaFactory');
 const recorder = require('./recorder');
-const { Given, When, Then } = require('./interfaces/context');
 
 let container = {
   helpers: {},
@@ -196,6 +195,7 @@ function getSupportObject(config, name) {
 }
 
 function loadGherkinSteps(paths) {
+  const { Given, When, Then } = require('./interfaces/bdd');
   for (const path of paths) {
     loadSupportObject(path, `Step Definition from ${path}`);
   }
