+++ lib/interfaces/gherkin.js
@@ -1,7 +1,8 @@
 const { Parser } = require('gherkin');
-const { matchStep } = require('./context');
+const { matchStep } = require('./bdd');
 const event = require('../event');
 const scenario = require('../scenario');
+const recorder = require('../recorder');
 const Step = require('../step');
 const Suite = require('mocha/lib/suite');
 const Context = require('mocha/lib/context');
@@ -43,21 +44,22 @@ module.exports = (text) => {
       continue;
     }
     if (child.type === 'ScenarioOutline') {
-      // console.log(child);
       for (const examples of child.examples) {
         const fields = examples.tableHeader.cells.map(c => c.value);
         for (const example of examples.tableBody) {
           let exampleSteps = [...child.steps];
+          const current = {};
           for (const index in example.cells) {
             const placeholder = fields[index];
             const value = example.cells[index].value;
+            current[placeholder] = value;
             exampleSteps = exampleSteps.map((step) => {
               step = Object.assign({}, step);
               step.text = step.text.replace(`<${placeholder}>`, value);
               return step;
             });
           }
-          const test = new Test(child.name, () => runSteps(exampleSteps));
+          const test = new Test(`${child.name} ${JSON.stringify(current)}`, () => runSteps(exampleSteps));
           test.timeout(0);
           test.async = true;
           suite.addTest(scenario.test(test));
