+++ lib/mochaFactory.js
@@ -5,7 +5,7 @@ const reporter = require('./reporter/cli');
 const gherkinParser = require('./interfaces/gherkin.js');
 const output = require('./output');
 
-const scenarioUi = fsPath.join(__dirname, './interfaces/bdd.js');
+const scenarioUi = fsPath.join(__dirname, './ui.js');
 
 let mocha;
