+++ lib/output.js
@@ -74,7 +74,13 @@ module.exports = {
     if (outputLevel === 0) return;
     if (!step) return;
     let stepLine = step.toString();
-    if (step.metaStep) stepLine = ' '.repeat(2) + colors.green(stepLine);
+    if (step.metaStep) {
+      if (outputLevel < 2) return;
+      stepLine = ' '.repeat(2) + colors.green(stepLine);
+    }
+    if (step.isMetaStep() && outputLevel === 2) {
+      stepLine += '\n'; // meta steps don't have execution time
+    }
     const sym = ' ';
     if (outputLevel === 2) {
       newline = false;
