+++ lib/step.js
@@ -98,6 +98,10 @@ class Step {
   toCode() {
     return `${this.prefix}${this.actor}.${this.name}(${this.humanizeArgs()})${this.suffix}`;
   }
+
+  isMetaStep() {
+    return this.constructor.name === 'MetaStep';
+  }
 }
